+++ gym_electric_motor/core.py
@@ -165,7 +165,7 @@ class ElectricMotorEnvironment(gym.core.Env):
         return self._visualizations
 
     def __init__(self, physical_system, reference_generator, reward_function, visualization=(), state_filter=None,
-                 callbacks=(), constraints=(), state_action_processors=(), **kwargs):
+                 callbacks=(), constraints=(), physical_system_wrappers=(), **kwargs):
         """
         Setting and initialization of all environments' modules.
 
@@ -184,7 +184,7 @@ class ElectricMotorEnvironment(gym.core.Env):
                         ConstraintMonitor in the environment.
             visualization(iterable(ElectricMotorVisualization)/None): The visualizations of this environment.
             state_filter(list(str)): Selection of states that are shown in the observation.
-            state_action_processors(iterable(StateActionProcessor)): StateActionProcessor instances to be wrapped around
+            physical_system_wrappers(iterable(PhysicalSystemWrapper)): PhysicalSystemWrapper instances to be wrapped around
                 the physical system.
             callbacks(list(Callback)): Callbacks being called in the environment
             **kwargs: Arguments to be passed to the modules.
@@ -207,8 +207,8 @@ class ElectricMotorEnvironment(gym.core.Env):
         self._constraint_monitor = cm
 
         # Announcement of the modules among each other
-        for state_action_processor in state_action_processors:
-            self._physical_system = state_action_processor.set_physical_system(self._physical_system)
+        for physical_system_wrapper in physical_system_wrappers:
+            self._physical_system = physical_system_wrapper.set_physical_system(self._physical_system)
         self._reference_generator.set_modules(self.physical_system)
         self._constraint_monitor.set_modules(self.physical_system)
         self._reward_function.set_modules(self.physical_system, self._reference_generator, self._constraint_monitor)
@@ -504,6 +504,9 @@ class PhysicalSystem:
 
     @property
     def unwrapped(self):
+        """Returns this instance of the physical system.
+        
+        If the system is wrapped into multiple PhysicalSystemWrappers this property returns directly the innermost system."""
         return self
 
     @property