+++ gym_electric_motor/envs/gym_dcm/permex_dc_motor_env/finite_sc_permex_dc_env.py
@@ -86,7 +86,7 @@ class FiniteSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
     def __init__(
         self, supply=None, converter=None, motor=None, load=None, ode_solver=None,
         reward_function=None, reference_generator=None, visualization=None, state_filter=None, callbacks=(),
-        constraints=('i',), calc_jacobian=True, tau=1e-5, state_action_processors=()
+        constraints=('i',), calc_jacobian=True, tau=1e-5, physical_system_wrappers=()
     ):
         """
         Args:
@@ -108,7 +108,7 @@ class FiniteSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
             tau(float): Duration of one control step in seconds. Default: 1e-5.
             state_filter(list(str)): List of states that shall be returned to the agent. Default: None (no filter)
             callbacks(list(Callback)): Callbacks for user interaction. Default: ()
-            state_action_processors(list(StateActionProcessor)): List of state action processors to modify the
+            physical_system_wrappers(list(PhysicalSystemWrapper)): List of Physical System Wrappers to modify the
             actions to and states from the physical system before they are used in the environment. Default: ()
 
         Note on the env-arg type:
@@ -148,5 +148,5 @@ class FiniteSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
         super().__init__(
             physical_system=physical_system, reference_generator=reference_generator, reward_function=reward_function,
             constraints=constraints, visualization=visualization, state_filter=state_filter, callbacks=callbacks,
-            state_action_processors=state_action_processors
+            physical_system_wrappers=physical_system_wrappers
         )