+++ gym_electric_motor/envs/gym_im/squirrel_cage_induction_motor_envs/cont_tc_scim_env.py
@@ -92,7 +92,7 @@ class ContTorqueControlSquirrelCageInductionMotorEnv(ElectricMotorEnvironment):
     def __init__(self, supply=None, converter=None, motor=None, load=None, ode_solver=None,
                  reward_function=None, reference_generator=None, visualization=None, state_filter=None, callbacks=(),
                  constraints=(SquaredConstraint(('i_sq', 'i_sd')),), calc_jacobian=True, tau=1e-4,
-                 state_action_processors=()):
+                 physical_system_wrappers=()):
         """
         Args:
             supply(env-arg): Specification of the :py:class:`.VoltageSupply` for the environment
@@ -113,7 +113,7 @@ class ContTorqueControlSquirrelCageInductionMotorEnv(ElectricMotorEnvironment):
             tau(float): Duration of one control step in seconds. Default: 1e-4.
             state_filter(list(str)): List of states that shall be returned to the agent. Default: None (no filter)
             callbacks(list(Callback)): Callbacks for user interaction. Default: ()
-            state_action_processors(list(StateActionProcessor)): List of state action processors to modify the
+            physical_system_wrappers(list(PhysicalSystemWrapper)): List of Physical System Wrappers to modify the
             actions to and states from the physical system before they are used in the environment. Default: ()
 
         Note on the env-arg type:
@@ -156,5 +156,5 @@ class ContTorqueControlSquirrelCageInductionMotorEnv(ElectricMotorEnvironment):
         super().__init__(
             physical_system=physical_system, reference_generator=reference_generator, reward_function=reward_function,
             constraints=constraints, visualization=visualization, state_filter=state_filter, callbacks=callbacks,
-            state_action_processors=state_action_processors
+            physical_system_wrappers=physical_system_wrappers
         )