+++ gym_electric_motor/physical_system_wrappers/current_sum_processor.py
@@ -1,10 +1,10 @@
 import gym
 import numpy as np
 
-from gym_electric_motor.state_action_processors import StateActionProcessor
+from gym_electric_motor.physical_system_wrappers import PhysicalSystemWrapper
 
 
-class CurrentSumProcessor(StateActionProcessor):
+class CurrentSumProcessor(PhysicalSystemWrapper):
     """Adds an ``i_sum`` state to the systems state vector that adds up currents."""
 
     def __init__(self, currents, limit='max', physical_system=None):