+++ gym_electric_motor/physical_systems/electric_motors/externally_excited_synchronous_motor.py
@@ -11,9 +11,10 @@ class ExternallyExcitedSynchronousMotor(SynchronousMotor):
     =====================  ==========  ============= ===========================================
     r_s                    mOhm        15.55         Stator resistance
     r_e                    mOhm        7.2           Excitation resistance
-    l_d                    mH          0.166         Direct axis inductance
+    l_d                    mH          1.66          Direct axis inductance
     l_q                    mH          0.35          Quadrature axis inductance
-    l_e                    mH          1.66          Excitation inductance
+    l_m                    mH          1.589         Mutual inductance
+    l_e                    mH          1.74          Excitation inductance
     p                      1           3             Pole pair number
     j_rotor                kg/m^2      0.3883        Moment of inertia of the rotor
     =====================  ==========  ============= ===========================================
@@ -98,8 +99,8 @@ class ExternallyExcitedSynchronousMotor(SynchronousMotor):
         'p': 3,
         'l_d': 1.66e-3,
         'l_q': 0.35e-3,
-        'l_m': 1.66e-4,
-        'l_e': 1.66e-3,
+        'l_m': 1.589e-3,
+        'l_e': 1.74e-3,
         'j_rotor': 0.3883,
         'r_s': 15.55e-3,
         'r_e': 7.2e-3,
@@ -161,16 +162,18 @@ class ExternallyExcitedSynchronousMotor(SynchronousMotor):
     def _torque_limit(self):
         # Docstring of superclass
         mp = self._motor_parameter
-        # I am not sure about this
         if mp['l_d'] == mp['l_q']:
             return self.torque([0, self._limits['i_sq'], self._limits['i_e'], 0])
         else:
             i_n = self.nominal_values['i']
             _p = mp['l_m'] * i_n / (2 * (mp['l_d'] - mp['l_q']))
             _q = - i_n ** 2 / 2
-            i_d_opt = - _p / 2 - np.sqrt( (_p / 2) ** 2 - _q)
+            if mp['l_d'] < mp['l_q']:
+                i_d_opt = - _p / 2 - np.sqrt( (_p / 2) ** 2 - _q)
+            else:
+                i_d_opt = - _p / 2 + np.sqrt( (_p / 2) ** 2 - _q)
             i_q_opt = np.sqrt(i_n ** 2 - i_d_opt ** 2)
-            return self.torque([i_d_opt, i_q_opt, self._limits['i_e'], 0])
+        return self.torque([i_d_opt, i_q_opt, self._limits['i_e'], 0])
 
     def torque(self, currents):
         # Docstring of superclass