+++ tests/utils/physical_system_test_wrapper.py
@@ -1,7 +1,7 @@
 import gym_electric_motor as gem
 
 
-class StateActionTestProcessor(gem.state_action_processors.StateActionProcessor):
+class PhysicalSystemTestWrapper(gem.physical_system_wrappers.PhysicalSystemWrapper):
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)